[mysql] Add various test cases to cover newly added table process from read binlog blocking to read binlog un-blocking
