+++ flink-connector-mysql-cdc/src/test/java/com/ververica/cdc/connectors/mysql/source/MySqlSourceITCase.java
@@ -21,12 +21,7 @@ import org.apache.flink.api.common.eventtime.WatermarkStrategy;
 import org.apache.flink.api.common.restartstrategy.RestartStrategies;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
-import org.apache.flink.configuration.Configuration;
 import org.apache.flink.core.execution.JobClient;
-import org.apache.flink.runtime.checkpoint.CheckpointException;
-import org.apache.flink.runtime.highavailability.nonha.embedded.HaLeadershipControl;
-import org.apache.flink.runtime.jobgraph.SavepointConfigOptions;
-import org.apache.flink.runtime.minicluster.MiniCluster;
 import org.apache.flink.streaming.api.datastream.DataStreamSource;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.api.operators.collect.CollectResultIterator;
@@ -40,7 +35,6 @@ import org.apache.flink.table.catalog.Column;
 import org.apache.flink.table.catalog.ResolvedSchema;
 import org.apache.flink.table.catalog.UniqueConstraint;
 import org.apache.flink.table.data.RowData;
-import org.apache.flink.table.planner.factories.TestValuesTableFactory;
 import org.apache.flink.table.runtime.typeutils.InternalTypeInfo;
 import org.apache.flink.table.types.DataType;
 import org.apache.flink.table.types.logical.LogicalType;
@@ -65,10 +59,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.kafka.connect.source.SourceRecord;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
 import org.junit.rules.Timeout;
 
-import java.lang.reflect.Field;
 import java.sql.SQLException;
 import java.time.Duration;
 import java.time.ZoneId;
@@ -80,7 +72,6 @@ import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Properties;
 import java.util.Random;
 import java.util.UUID;
@@ -239,119 +230,6 @@ public class MySqlSourceITCase extends MySqlSourceTestBase {
     public void testJobManagerFailoverSingleParallelism() throws Exception {
         testMySqlParallelSource(
                 1, FailoverType.JM, FailoverPhase.SNAPSHOT, new String[] {"customers"});
-    }
-
-    @Test
-    public void testNewlyAddedTableForExistsPipelineOnce() throws Exception {
-        testNewlyAddedTableOneByOne(
-                1,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                false,
-                "address_hangzhou",
-                "address_beijing");
-    }
-
-    @Test
-    public void testNewlyAddedTableForExistsPipelineOnceWithAheadBinlog() throws Exception {
-        testNewlyAddedTableOneByOne(
-                1,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                true,
-                "address_hangzhou",
-                "address_beijing");
-    }
-
-    @Test
-    public void testNewlyAddedTableForExistsPipelineTwice() throws Exception {
-        testNewlyAddedTableOneByOne(
-                DEFAULT_PARALLELISM,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                false,
-                "address_hangzhou",
-                "address_beijing",
-                "address_shanghai");
-    }
-
-    @Test
-    public void testNewlyAddedTableForExistsPipelineTwiceWithAheadBinlog() throws Exception {
-        testNewlyAddedTableOneByOne(
-                DEFAULT_PARALLELISM,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                true,
-                "address_hangzhou",
-                "address_beijing",
-                "address_shanghai");
-    }
-
-    @Test
-    public void testNewlyAddedTableForExistsPipelineSingleParallelism() throws Exception {
-        testNewlyAddedTableOneByOne(
-                1,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                false,
-                "address_hangzhou",
-                "address_beijing");
-    }
-
-    @Test
-    public void testNewlyAddedTableForExistsPipelineSingleParallelismWithAheadBinlog()
-            throws Exception {
-        testNewlyAddedTableOneByOne(
-                1,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                true,
-                "address_hangzhou",
-                "address_beijing");
-    }
-
-    @Test
-    public void testJobManagerFailoverForNewlyAddedTable() throws Exception {
-        testNewlyAddedTableOneByOne(
-                DEFAULT_PARALLELISM,
-                FailoverType.JM,
-                FailoverPhase.SNAPSHOT,
-                false,
-                "address_hangzhou",
-                "address_beijing");
-    }
-
-    @Test
-    public void testJobManagerFailoverForNewlyAddedTableWithAheadBinlog() throws Exception {
-        testNewlyAddedTableOneByOne(
-                DEFAULT_PARALLELISM,
-                FailoverType.JM,
-                FailoverPhase.SNAPSHOT,
-                true,
-                "address_hangzhou",
-                "address_beijing");
-    }
-
-    @Test
-    public void testTaskManagerFailoverForNewlyAddedTable() throws Exception {
-        testNewlyAddedTableOneByOne(
-                1,
-                FailoverType.TM,
-                FailoverPhase.BINLOG,
-                false,
-                "address_hangzhou",
-                "address_beijing");
-    }
-
-    @Test
-    public void testTaskManagerFailoverForNewlyAddedTableWithAheadBinlog() throws Exception {
-        testNewlyAddedTableOneByOne(
-                1,
-                FailoverType.TM,
-                FailoverPhase.BINLOG,
-                false,
-                "address_hangzhou",
-                "address_beijing");
     }
 
     @Test
@@ -756,207 +634,6 @@ public class MySqlSourceITCase extends MySqlSourceTestBase {
                 .collect(Collectors.toList());
     }
 
-    private void testNewlyAddedTableOneByOne(
-            int parallelism,
-            FailoverType failoverType,
-            FailoverPhase failoverPhase,
-            boolean makeBinlogBeforeCapture,
-            String... captureAddressTables)
-            throws Exception {
-
-        // step 1: create mysql tables with initial data
-        TestValuesTableFactory.clearAllData();
-        customDatabase.createAndInitialize();
-        initialAddressTables(getConnection(), captureAddressTables);
-
-        final TemporaryFolder temporaryFolder = new TemporaryFolder();
-        temporaryFolder.create();
-        final String savepointDirectory = temporaryFolder.newFolder().toURI().toString();
-
-        // test newly added table one by one
-        String finishedSavePointPath = null;
-        List<String> fetchedDataList = new ArrayList<>();
-        for (int round = 0; round < captureAddressTables.length; round++) {
-            String[] captureTablesThisRound =
-                    Arrays.asList(captureAddressTables)
-                            .subList(0, round + 1)
-                            .toArray(new String[0]);
-            String newlyAddedTable = captureAddressTables[round];
-            if (makeBinlogBeforeCapture) {
-                makeBinlogBeforeCaptureForAddressTable(getConnection(), newlyAddedTable);
-            }
-            StreamExecutionEnvironment env =
-                    getStreamExecutionEnvironment(finishedSavePointPath, parallelism);
-            StreamTableEnvironment tEnv = StreamTableEnvironment.create(env);
-
-            String createTableStatement = getCreateTableStatement(captureTablesThisRound);
-            tEnv.executeSql(createTableStatement);
-            tEnv.executeSql(
-                    "CREATE TABLE sink ("
-                            + " table_name STRING,"
-                            + " id BIGINT,"
-                            + " country STRING,"
-                            + " city STRING,"
-                            + " detail_address STRING,"
-                            + " primary key (id) not enforced"
-                            + ") WITH ("
-                            + " 'connector' = 'values',"
-                            + " 'sink-insert-only' = 'false'"
-                            + ")");
-            TableResult tableResult = tEnv.executeSql("insert into sink select * from address");
-            JobClient jobClient = tableResult.getJobClient().get();
-
-            // step 2: assert fetched snapshot data in this round
-            String cityName = newlyAddedTable.split("_")[1];
-            List<String> expectedSnapshotDataThisRound =
-                    Arrays.asList(
-                            format(
-                                    "+I[%s, 416874195632735147, China, %s, %s West Town address 1]",
-                                    newlyAddedTable, cityName, cityName),
-                            format(
-                                    "+I[%s, 416927583791428523, China, %s, %s West Town address 2]",
-                                    newlyAddedTable, cityName, cityName),
-                            format(
-                                    "+I[%s, 417022095255614379, China, %s, %s West Town address 3]",
-                                    newlyAddedTable, cityName, cityName));
-            if (makeBinlogBeforeCapture) {
-                expectedSnapshotDataThisRound =
-                        Arrays.asList(
-                                format(
-                                        "+I[%s, 416874195632735147, China, %s, %s West Town address 1]",
-                                        newlyAddedTable, cityName, cityName),
-                                format(
-                                        "+I[%s, 416927583791428523, China, %s, %s West Town address 2]",
-                                        newlyAddedTable, cityName, cityName),
-                                format(
-                                        "+I[%s, 417022095255614379, China, %s, %s West Town address 3]",
-                                        newlyAddedTable, cityName, cityName),
-                                format(
-                                        "+I[%s, 417022095255614381, China, %s, %s West Town address 5]",
-                                        newlyAddedTable, cityName, cityName));
-            }
-
-            // trigger failover after some snapshot data read finished
-            if (failoverPhase == FailoverPhase.SNAPSHOT) {
-                triggerFailover(
-                        failoverType,
-                        jobClient.getJobID(),
-                        miniClusterResource.getMiniCluster(),
-                        () -> sleepMs(100));
-            }
-            fetchedDataList.addAll(expectedSnapshotDataThisRound);
-            waitForSinkSize("sink", fetchedDataList.size());
-            assertEqualsInAnyOrder(fetchedDataList, TestValuesTableFactory.getRawResults("sink"));
-
-            // step 3: make some binlog data for this round
-            makeFirstPartBinlogForAddressTable(getConnection(), newlyAddedTable);
-            if (failoverPhase == FailoverPhase.BINLOG) {
-                triggerFailover(
-                        failoverType,
-                        jobClient.getJobID(),
-                        miniClusterResource.getMiniCluster(),
-                        () -> sleepMs(100));
-            }
-            makeSecondPartBinlogForAddressTable(getConnection(), newlyAddedTable);
-
-            // step 4: assert fetched binlog data in this round
-            List<String> expectedBinlogDataThisRound =
-                    Arrays.asList(
-                            format(
-                                    "+U[%s, 416874195632735147, CHINA, %s, %s West Town address 1]",
-                                    newlyAddedTable, cityName, cityName),
-                            format(
-                                    "+I[%s, 417022095255614380, China, %s, %s West Town address 4]",
-                                    newlyAddedTable, cityName, cityName));
-
-            // step 5: assert fetched binlog data in this round
-            fetchedDataList.addAll(expectedBinlogDataThisRound);
-            waitForSinkSize("sink", fetchedDataList.size());
-            assertEqualsInAnyOrder(fetchedDataList, TestValuesTableFactory.getRawResults("sink"));
-
-            // step 6: trigger savepoint
-            if (round != captureAddressTables.length - 1) {
-                finishedSavePointPath = triggerSavepointWithRetry(jobClient, savepointDirectory);
-            }
-            jobClient.cancel().get();
-        }
-    }
-
-    private String getCreateTableStatement(String... captureTableNames) {
-        return format(
-                "CREATE TABLE address ("
-                        + " table_name STRING METADATA VIRTUAL,"
-                        + " id BIGINT NOT NULL,"
-                        + " country STRING,"
-                        + " city STRING,"
-                        + " detail_address STRING,"
-                        + " primary key (id) not enforced"
-                        + ") WITH ("
-                        + " 'connector' = 'mysql-cdc',"
-                        + " 'scan.incremental.snapshot.enabled' = 'true',"
-                        + " 'hostname' = '%s',"
-                        + " 'port' = '%s',"
-                        + " 'username' = '%s',"
-                        + " 'password' = '%s',"
-                        + " 'database-name' = '%s',"
-                        + " 'table-name' = '%s',"
-                        + " 'scan.incremental.snapshot.chunk.size' = '2',"
-                        + " 'server-time-zone' = 'UTC',"
-                        + " 'server-id' = '%s',"
-                        + " 'scan.newly-added-table.enabled' = 'true'"
-                        + ")",
-                MYSQL_CONTAINER.getHost(),
-                MYSQL_CONTAINER.getDatabasePort(),
-                customDatabase.getUsername(),
-                customDatabase.getPassword(),
-                customDatabase.getDatabaseName(),
-                getTableNameRegex(captureTableNames),
-                getServerId());
-    }
-
-    private StreamExecutionEnvironment getStreamExecutionEnvironment(
-            String finishedSavePointPath, int parallelism) throws Exception {
-        StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
-        if (finishedSavePointPath != null) {
-            // restore from savepoint
-            // hack for test to visit protected TestStreamEnvironment#getConfiguration() method
-            ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-            Class<?> clazz =
-                    classLoader.loadClass(
-                            "org.apache.flink.streaming.api.environment.StreamExecutionEnvironment");
-            Field field = clazz.getDeclaredField("configuration");
-            field.setAccessible(true);
-            Configuration configuration = (Configuration) field.get(env);
-            configuration.setString(SavepointConfigOptions.SAVEPOINT_PATH, finishedSavePointPath);
-        }
-        env.setParallelism(parallelism);
-        env.enableCheckpointing(200L);
-        env.setRestartStrategy(RestartStrategies.fixedDelayRestart(3, 100L));
-        return env;
-    }
-
-    private String triggerSavepointWithRetry(JobClient jobClient, String savepointDirectory)
-            throws ExecutionException, InterruptedException {
-        int retryTimes = 0;
-        // retry 600 times, it takes 100 milliseconds per time, at most retry 1 minute
-        while (retryTimes < 600) {
-            try {
-                return jobClient.triggerSavepoint(savepointDirectory).get();
-            } catch (Exception e) {
-                Optional<CheckpointException> exception =
-                        ExceptionUtils.findThrowable(e, CheckpointException.class);
-                if (exception.isPresent()
-                        && exception.get().getMessage().contains("Checkpoint triggering task")) {
-                    Thread.sleep(100);
-                    retryTimes++;
-                } else {
-                    throw e;
-                }
-            }
-        }
-        return null;
-    }
-
     private static List<String> fetchRows(Iterator<Row> iter, int size) {
         List<String> rows = new ArrayList<>(size);
         while (size > 0 && iter.hasNext()) {
@@ -1047,89 +724,6 @@ public class MySqlSourceITCase extends MySqlSourceTestBase {
         }
     }
 
-    private void initialAddressTables(JdbcConnection connection, String[] addressTables)
-            throws SQLException {
-        try {
-            connection.setAutoCommit(false);
-            for (String tableName : addressTables) {
-                // make initial data for given table
-                String tableId = customDatabase.getDatabaseName() + "." + tableName;
-                String cityName = tableName.split("_")[1];
-                connection.execute(
-                        "CREATE TABLE "
-                                + tableId
-                                + "("
-                                + "  id BIGINT UNSIGNED NOT NULL PRIMARY KEY,"
-                                + "  country VARCHAR(255) NOT NULL,"
-                                + "  city VARCHAR(255) NOT NULL,"
-                                + "  detail_address VARCHAR(1024)"
-                                + ");");
-                connection.execute(
-                        format(
-                                "INSERT INTO  %s "
-                                        + "VALUES (416874195632735147, 'China', '%s', '%s West Town address 1'),"
-                                        + "       (416927583791428523, 'China', '%s', '%s West Town address 2'),"
-                                        + "       (417022095255614379, 'China', '%s', '%s West Town address 3');",
-                                tableId, cityName, cityName, cityName, cityName, cityName,
-                                cityName));
-            }
-            connection.commit();
-        } finally {
-            connection.close();
-        }
-    }
-
-    private void makeFirstPartBinlogForAddressTable(JdbcConnection connection, String tableName)
-            throws SQLException {
-        try {
-            connection.setAutoCommit(false);
-            // make binlog events for the first split
-            String tableId = customDatabase.getDatabaseName() + "." + tableName;
-            String cityName = tableName.split("_")[1];
-            connection.execute(
-                    format(
-                            "UPDATE %s SET COUNTRY = 'CHINA' where id = 416874195632735147",
-                            tableId));
-            connection.commit();
-        } finally {
-            connection.close();
-        }
-    }
-
-    private void makeSecondPartBinlogForAddressTable(JdbcConnection connection, String tableName)
-            throws SQLException {
-        try {
-            connection.setAutoCommit(false);
-            // make binlog events for the second split
-            String tableId = customDatabase.getDatabaseName() + "." + tableName;
-            String cityName = tableName.split("_")[1];
-            connection.execute(
-                    format(
-                            "INSERT INTO %s VALUES(417022095255614380, 'China','%s','%s West Town address 4')",
-                            tableId, cityName, cityName));
-            connection.commit();
-        } finally {
-            connection.close();
-        }
-    }
-
-    private void makeBinlogBeforeCaptureForAddressTable(JdbcConnection connection, String tableName)
-            throws SQLException {
-        try {
-            connection.setAutoCommit(false);
-            // make binlog before the capture of the table
-            String tableId = customDatabase.getDatabaseName() + "." + tableName;
-            String cityName = tableName.split("_")[1];
-            connection.execute(
-                    format(
-                            "INSERT INTO %s VALUES(417022095255614381, 'China','%s','%s West Town address 5')",
-                            tableId, cityName, cityName));
-            connection.commit();
-        } finally {
-            connection.close();
-        }
-    }
-
     private MySqlConnection getConnection() {
         Map<String, String> properties = new HashMap<>();
         properties.put("database.hostname", MYSQL_CONTAINER.getHost());
@@ -1140,74 +734,6 @@ public class MySqlSourceITCase extends MySqlSourceTestBase {
         io.debezium.config.Configuration configuration =
                 io.debezium.config.Configuration.from(properties);
         return DebeziumUtils.createMySqlConnection(configuration, new Properties());
-    }
-
-    // ------------------------------------------------------------------------
-    //  test utilities
-    // ------------------------------------------------------------------------
-
-    /** The type of failover. */
-    private enum FailoverType {
-        TM,
-        JM,
-        NONE
-    }
-
-    /** The phase of failover. */
-    private enum FailoverPhase {
-        SNAPSHOT,
-        BINLOG,
-        NEVER
-    }
-
-    private static void triggerFailover(
-            FailoverType type, JobID jobId, MiniCluster miniCluster, Runnable afterFailAction)
-            throws Exception {
-        switch (type) {
-            case TM:
-                restartTaskManager(miniCluster, afterFailAction);
-                break;
-            case JM:
-                triggerJobManagerFailover(jobId, miniCluster, afterFailAction);
-                break;
-            case NONE:
-                break;
-            default:
-                throw new IllegalStateException("Unexpected value: " + type);
-        }
-    }
-
-    private static void triggerJobManagerFailover(
-            JobID jobId, MiniCluster miniCluster, Runnable afterFailAction) throws Exception {
-        final HaLeadershipControl haLeadershipControl = miniCluster.getHaLeadershipControl().get();
-        haLeadershipControl.revokeJobMasterLeadership(jobId).get();
-        afterFailAction.run();
-        haLeadershipControl.grantJobMasterLeadership(jobId).get();
-    }
-
-    private static void restartTaskManager(MiniCluster miniCluster, Runnable afterFailAction)
-            throws Exception {
-        miniCluster.terminateTaskManager(0).get();
-        afterFailAction.run();
-        miniCluster.startTaskManager();
-    }
-
-    private static void waitForSinkSize(String sinkName, int expectedSize)
-            throws InterruptedException {
-        while (sinkSize(sinkName) < expectedSize) {
-            Thread.sleep(100);
-        }
-    }
-
-    private static int sinkSize(String sinkName) {
-        synchronized (TestValuesTableFactory.class) {
-            try {
-                return TestValuesTableFactory.getRawResults(sinkName).size();
-            } catch (IllegalArgumentException e) {
-                // job is not started yet
-                return 0;
-            }
-        }
     }
 
     private void waitUntilJobRunning(TableResult tableResult)
